Moving the workflow printing to separate object
1. Modified models to remove the default value for seed phrase
2. Move the workflow printing to separate object just to make it optional
3. Adding hdDerivePath to the vault data
