+++ shared/src/main/scala/com/livelygig/product/shared/models/wallet/WalletModels.scala
@@ -35,7 +35,7 @@ object WalletDetails {
   implicit val format = Json.format[WalletDetails]
 }
 
-case class UserDetails(alias: String, walletDetails: WalletDetails)
+case class UserDetails(alias: String, walletDetails: WalletDetails, phrase: Seq[String])
 
 object UserDetails {
   implicit val format = Json.format[UserDetails]
