+++ wallet-client/src/main/scala/com/livelygig/product/walletclient/views/AccountView.scala
@@ -31,7 +31,7 @@ object AccountView {
   Toastr.options.positionClass = "toast-top-full-width"
   case class Props(proxy: ModelProxy[Pot[ERCTokenRootModel]], router: RouterCtl[Loc])
 
-  final case class State(currencySelected: String, coinExchange: CoinExchange, userDetails: UserDetails = UserDetails("", WalletDetails("", "")))
+  final case class State(currencySelected: String, coinExchange: CoinExchange, userDetails: UserDetails = UserDetails("", WalletDetails("", ""), Nil))
 
   final class Backend(t: BackendScope[Props, State]) {
     def getLiveCurrencyUpdate() = {
@@ -202,7 +202,7 @@ object AccountView {
   }
 
   val component = ScalaComponent.builder[Props]("AccountView")
-    .initialState(State("ETH", CoinExchange(Seq(CurrencyList("", Seq(Currency("", 0, ""))))), UserDetails("", WalletDetails("", "0"))))
+    .initialState(State("ETH", CoinExchange(Seq(CurrencyList("", Seq(Currency("", 0, ""))))), UserDetails("", WalletDetails("", "0"), Nil)))
     .renderBackend[Backend]
     .componentWillMount(scope => scope.backend.updateCurrency())
     .componentDidMount(scope => scope.backend.componentDidMount(scope.props))
