+++ wallet-client/src/main/scala/com/livelygig/product/walletclient/views/ViewBackupPhrase.scala
@@ -19,7 +19,6 @@ object ViewBackupPhrase {
   final class Backend(t: BackendScope[Props, State]) {
 
     def componentDidMount(props: Props): Callback = {
-//      WalletCircuit.dispatch(GetUserDetails(WalletCircuit.zoom(_.user).value.userDetails.copy(phrase = "")))
       Callback.empty
     }
